Move `rebuild_cache` to the codegen module
